Move DAO logic to new file. Parameterize DAO tests
